+++ src/main/java/co/edu/udistrital/mdp/adopcion/controllers/events/ShelterArrivalController.java
@@ -9,6 +9,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
@@ -16,13 +19,9 @@ import org.springframework.web.bind.annotation.RestController;
 import co.edu.udistrital.mdp.adopcion.dto.events.ShelterArrivalDTO;
 import co.edu.udistrital.mdp.adopcion.dto.events.ShelterArrivalDetailDTO;
 import co.edu.udistrital.mdp.adopcion.entities.events.ShelterArrivalEntity;
-import co.edu.udistrital.mdp.adopcion.services.events.ShelterArrivalService;
-
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-
 import co.edu.udistrital.mdp.adopcion.exceptions.EntityNotFoundException;
 import co.edu.udistrital.mdp.adopcion.exceptions.IllegalOperationException;
+import co.edu.udistrital.mdp.adopcion.services.events.ShelterArrivalService;
 
 
 @RestController
@@ -52,11 +51,15 @@ public class ShelterArrivalController {
      * 
      * @param shelter_arrival_id
      * @return
+     * @throws EntityNotFoundException
      */
     @GetMapping(value = "/{shelter_arrival_id}")
     @ResponseStatus(code = HttpStatus.OK)
-    public ShelterArrivalDetailDTO findOne(@PathVariable Long shelter_arrival_id) {
+    public ShelterArrivalDetailDTO findOne(@PathVariable Long shelter_arrival_id) throws EntityNotFoundException {
         ShelterArrivalEntity shelterArrivalEntity = shelterArrivalService.getShelterArrivalById(shelter_arrival_id);
+        if (shelterArrivalEntity == null) {
+            throw new EntityNotFoundException("Shelter arrival not found with ID: " + shelter_arrival_id);
+        }
         return modelMapper.map(shelterArrivalEntity, ShelterArrivalDetailDTO.class);
     }
 
@@ -80,11 +83,14 @@ public class ShelterArrivalController {
      * @param shelterArrivalDTO
      * @return
      */
-    @PostMapping(value = "/{shelter_arrival_id}")
+    @PutMapping(value = "/{shelter_arrival_id}")
     @ResponseStatus(code = HttpStatus.OK)
     public ShelterArrivalDTO update(@PathVariable Long shelter_arrival_id, @RequestBody ShelterArrivalDTO shelterArrivalDTO) 
             throws EntityNotFoundException, IllegalOperationException {
         ShelterArrivalEntity shelterArrivalEntity = shelterArrivalService.updateShelterArrival(shelter_arrival_id, modelMapper.map(shelterArrivalDTO, ShelterArrivalEntity.class));
+        if (shelterArrivalEntity == null) {
+            throw new EntityNotFoundException("Shelter arrival not found with ID: " + shelter_arrival_id);
+        }
         return modelMapper.map(shelterArrivalEntity, ShelterArrivalDTO.class);
     }
 
@@ -96,6 +102,9 @@ public class ShelterArrivalController {
     @DeleteMapping(value = "/{shelter_arrival_id}")
     @ResponseStatus(code = HttpStatus.NO_CONTENT)
     public void delete(@PathVariable Long shelter_arrival_id) throws EntityNotFoundException, IllegalOperationException {
+        if (shelterArrivalService.getShelterArrivalById(shelter_arrival_id) == null) {
+            throw new EntityNotFoundException("Shelter arrival not found with ID: " + shelter_arrival_id);
+        }
         shelterArrivalService.deleteShelterArrival(shelter_arrival_id);
     }
